add mock data and make seed work small adjustments at schema.prisma with franz
