--- assets/data.ts
@@ -1,193 +1,0 @@
-export const mockData = [
-  {
-    firstname: "Johannes",
-    lastname: "Smith",
-    email: "johannes.smith@email.de",
-    passwordHash: "effegysdgerzerz",
-    passwordSalt: "bnfhjfdjdfjgfj",
-    profilePicture: "https://unsplash.com/photos/OhKElOkQ3RE",
-    rating: 5,
-    item: {
-      title: "Shoe",
-      images: "https://unsplash.com/photos/164_6wVEHfI",
-      description: "mint condition, unused",
-      sellType: "FREE",
-      category: "Clothes",
-    },
-    favorite: "0",
-    location: {
-      lat: 51.33573247467917,
-      lon: 12.408345798001886,
-      address: "Beuchaer Str. 2, 04318 Leipzig",
-    },
-    conversations: [],
-  },
-
-  {
-    firstname: "Rolf",
-    lastname: "Müller",
-    email: "rolf.müller@email.de",
-    passwordHash: "asfhdfghfd",
-    passwordSalt: "bnfhjfgdfgfddjdfjgfj",
-    profilePicture: "https://unsplash.com/photos/YUu9UAcOKZ4",
-    rating: 2,
-    item: {
-      title: "Hat",
-      images: "https://unsplash.com/photos/rr4bawLxOjc",
-      description: "Classic, blue hat",
-      sellType: "SWAP",
-      category: "Clothes",
-    },
-    favorite: "2",
-    location: {
-      lat: 51.337493607362845,
-      lon: 12.40253711886957,
-      address: "Anna-Kuhnow-Straße 16, 04317 Leipzig",
-    },
-    conversations: [],
-  },
-
-  {
-    firstname: "Peter",
-    lastname: "Winston",
-    email: "peter.winston@email.de",
-    passwordHash: "kuziktz",
-    passwordSalt: "aewrsrse",
-    profilePicture: "https://unsplash.com/photos/iFgRcqHznqg",
-    rating: 1,
-    item: {
-      title: "Boot",
-      images: "https://unsplash.com/photos/esxf7PJmExQ",
-      description: "Bad condition, strongly used",
-      sellType: "FREE",
-      category: "Clothes",
-    },
-
-    favorite: "0",
-    location: {
-      lat: 51.336718632668116,
-      lon: 12.406256671762284,
-      address: "Kippenbergstraße 28, 04317 Leipzig",
-    },
-    conversations: [],
-  },
-
-  {
-    firstname: "Petra",
-    lastname: "Hurz",
-    email: "petra.hurz@email.de",
-    passwordHash: "ögfkgf",
-    passwordSalt: "asfhfd",
-    profilePicture: "https://unsplash.com/photos/mEZ3PoFGs_k",
-    rating: 2,
-    item: {
-      title: "Glasses",
-      images: "https://unsplash.com/photos/WeCoLo0Rxp4",
-      description: "One glass is broken, but otherwise good",
-      sellType: "SWAP",
-      category: "Clothes",
-    },
-    favorite: "5",
-    location: {
-      lat: 51.3357937256552,
-      lon: 12.406558419972988,
-      address: "Täubchenweg 86, 04317 Leipzig",
-    },
-    conversations: [],
-  },
-
-  {
-    firstname: "Simone",
-    lastname: "Malz",
-    email: "simone.malz@email.de",
-    passwordHash: "kogspdfosd",
-    passwordSalt: "öojgpodsjgsd",
-    profilePicture: "https://unsplash.com/photos/AJIqZDAUD7A",
-    rating: 4,
-    item: {
-      title: "Tie",
-      images: "https://unsplash.com/photos/oZFqTBuQw_k",
-      description: "nice animals, only used at wedding",
-      sellType: "SWAP",
-      category: "Clothes",
-    },
-    favorite: "2",
-    location: {
-      lat: 51.33959900523878,
-      lon: 12.40713727722273,
-      address: "Lilienstraße 39a, 04315 Leipzig",
-    },
-    conversations: [],
-  },
-
-  {
-    firstname: "Ines",
-    lastname: "Schwarz",
-    email: "ines.schwarz@email.de",
-    passwordHash: "gsdgds",
-    passwordSalt: "AFHFD",
-    profilePicture: "https://unsplash.com/photos/rDEOVtE7vOs",
-    rating: 5,
-    item: {
-      title: "Blouse",
-      images: "https://unsplash.com/photos/LLHs1CN-5UU",
-      description: "white, classic",
-      sellType: "FREE",
-      category: "Clothes",
-    },
-    favorite: "1",
-    location: {
-      lat: 51.33799145756473,
-      lon: 12.401286770922855,
-      address: "Heinrichstraße 16, 04317 Leipzig",
-    },
-    conversations: [],
-  },
-
-  {
-    firstname: "Alistair",
-    lastname: "Fisher",
-    email: "alistair.fisher@email.de",
-    passwordHash: "gdfgdf",
-    passwordSalt: "gdgdf",
-    profilePicture: "https://unsplash.com/photos/7YVZYZeITc8",
-    rating: 2,
-    item: {
-      title: "Suit",
-      images: "https://unsplash.com/photos/ECJWF1NxSpM",
-      description: "suit including shirt, vest, jacket, trousers and tie",
-      sellType: "FREE",
-      category: "Clothes",
-    },
-    favorite: "4",
-    location: {
-      lat: 51.339674129638844,
-      lon: 12.410080043100086,
-      address: "Bernhardstraße 27, 04315 Leipzig",
-    },
-    conversations: [],
-  },
-  {
-    firstname: "Aylin",
-    lastname: "Yilmaz",
-    email: "aylin.yilmaz@email.de",
-    passwordHash: "khggb",
-    passwordSalt: "sgjzrj",
-    profilePicture: "https://unsplash.com/photos/W7b3eDUb_2I",
-    rating: 2,
-    item: {
-      title: "Sweatshirt",
-      images: "https://unsplash.com/photos/3G1TJktMQJw",
-      description: "not in a really good condition, black",
-      sellType: "FREE",
-      category: "Clothes",
-    },
-    favorite: "1",
-    location: {
-      lat: 51.3413036318854,
-      lon: 12.406251931287448,
-      address: "Hermann-Liebmann-Straße 27, 04315 Leipzig",
-    },
-    conversations: [],
-  },
-];
