+++ package.json
@@ -12,9 +12,9 @@
     "lint": "next lint",
     "storybook": "start-storybook -p 6006",
     "build-storybook": "build-storybook",
-    "start:prisma": "npx prismix && npx prisma format && npx prisma migrate dev && ts-node src/server.ts",
+    "start:prisma": "npx prisma format && npx prisma migrate dev && ts-node src/server.ts",
     "db:push": "npx prisma db push",
-    "db:migrate": "npx prismix && npx prisma format && npx prisma migrate dev",
+    "db:migrate": "npx prisma format && npx prisma migrate dev",
     "db:studio": "npx prisma studio",
     "test": "jest",
     "test:e2e": "playwright test",
